refactor code Day4_Bai1 and add fix feature pause_Game
