+++ Chiet-08-K5/assets/Day3/Bai1/Data_31/Data31.js
@@ -1,0 +1,35 @@
+// const Item = require('Item');
+
+// cc.Class({
+//     name: 'Data',
+//     properties: {
+//         inventory: []
+//     },
+
+//     ctor() {
+//         this.inventory = [
+//             new Item("Bình máu", 3, "consumable", "Hồi 20 máu"),
+//             new Item("Kiếm Sắt", 1, "equipment", "Tăng 10 sức mạnh"),
+//             new Item("Bình năng lượng", 2, "consumable", "Hồi 15 năng lượng"),
+//             new Item("Giáp Sắt", 1, "equipment", "Tăng 20 giáp"),
+//             new Item("Bình máu lớn", 1, "consumable", "Hồi 50 máu"),
+//             new Item("Kiếm Lửa", 1, "equipment", "Tăng 25 sức mạnh và gây sát thương lửa"),
+//             new Item("Khiên Gỗ", 1, "equipment", "Tăng 10 giáp"),
+//             new Item("Bình độc", 2, "consumable", "Gây 30 sát thương cho kẻ địch"),
+//             new Item("Vòng cổ hồi phục", 1, "equipment", "Hồi 5 máu mỗi lượt"),
+//             new Item("Thuốc tàng hình", 1, "consumable", "Tàng hình trong 10 giây")
+//         ];
+//     },
+
+//     getItemList() {
+//         return this.inventory;
+//     },
+
+//     getItem(index) {
+//         return this.inventory[index];
+//     },
+
+//     removeItem(index) {
+//         this.inventory.splice(index, 1);
+//     }
+// });
