+++ Chiet-08-K5/assets/Day3/Bai1/Data_31/MainScript31.js
@@ -1,44 +1,44 @@
-const Data = require('Data');
-const GameController = require('GameController'); 
+// const Data = require('Data');
+// const GameController = require('GameController');
 
-cc.Class({
-    extends: cc.Component,
+// cc.Class({
+//     extends: cc.Component,
 
-    properties: {
-        itemPrefab: cc.Prefab,
-        itemListNode: cc.Node,
-        nameLabel: cc.Label,
-        quantityLabel: cc.Label,
-        typeLabel: cc.Label,
-        effectLabel: cc.Label,
-        useButton: cc.Button,
-        deleteButton: cc.Button,
-        messageLabel: cc.Label
-    },
+//     properties: {
+//         itemPrefab: cc.Prefab,
+//         itemListNode: cc.Node,
+//         nameLabel: cc.Label,
+//         quantityLabel: cc.Label,
+//         typeLabel: cc.Label,
+//         effectLabel: cc.Label,
+//         useButton: cc.Button,
+//         deleteButton: cc.Button,
+//         messageLabel: cc.Label
+//     },
 
-    onLoad() {
-        const data = new Data();
-        this.gameController = new GameController(data, this);
-        this.gameController.showItem();
-    },
+//     onLoad() {
+//         const data = new Data();
+//         this.gameController = new GameController(data, this);
+//         this.gameController.showItem();
+//     },
 
-    useItem() {
-        this.gameController.useItem();
-    },
+//     useItem() {
+//         this.gameController.useItem();
+//     },
 
-    deleteItem() {
-        this.gameController.deleteItem();
-    },
+//     deleteItem() {
+//         this.gameController.deleteItem();
+//     },
 
-    showMessage(msg) {
-        this.messageLabel.string = msg;
-        this.messageLabel.node.active = true;
+//     showMessage(msg) {
+//         this.messageLabel.string = msg;
+//         this.messageLabel.node.active = true;
 
-        this.unschedule(this._hideMessage);
-        this.scheduleOnce(this._hideMessage, 2);
-    },
+//         this.unschedule(this._hideMessage);
+//         this.scheduleOnce(this._hideMessage, 2);
+//     },
 
-    _hideMessage() {
-        this.messageLabel.node.active = false;
-    }
-});
+//     _hideMessage() {
+//         this.messageLabel.node.active = false;
+//     }
+// });
