+++ Chiet-08-K5/assets/Day4/Bai1/Data_41/Prefab.meta
@@ -1,0 +1,7 @@
+{
+  "ver": "1.0.1",
+  "uuid": "0fa18836-a3a7-4315-a172-d710df24fda6",
+  "isSubpackage": false,
+  "subpackageName": "",
+  "subMetas": {}
+}
