+++ Chiet-08-K5/assets/Day4/Bai1/Data_41/Scene.meta
@@ -1,0 +1,7 @@
+{
+  "ver": "1.0.1",
+  "uuid": "df1e0740-292d-4910-be00-818787388b04",
+  "isSubpackage": false,
+  "subpackageName": "",
+  "subMetas": {}
+}
