+++ Chiet-08-K5/assets/Day4/Bai1/Data_41/Scripts.meta
@@ -1,0 +1,7 @@
+{
+  "ver": "1.0.1",
+  "uuid": "749395a9-59ce-4046-9fe6-20a0da4bf838",
+  "isSubpackage": false,
+  "subpackageName": "",
+  "subMetas": {}
+}
