+++ Chiet-08-K5/assets/images/PauseButton.png.meta
@@ -1,0 +1,34 @@
+{
+  "ver": "2.3.3",
+  "uuid": "96b0e8ee-ba50-404e-8010-58f52bc7cc5d",
+  "type": "sprite",
+  "wrapMode": "clamp",
+  "filterMode": "bilinear",
+  "premultiplyAlpha": false,
+  "genMipmaps": false,
+  "packable": true,
+  "platformSettings": {},
+  "subMetas": {
+    "PauseButton": {
+      "ver": "1.0.4",
+      "uuid": "0743d8de-af15-4869-a108-4c2707939301",
+      "rawTextureUuid": "96b0e8ee-ba50-404e-8010-58f52bc7cc5d",
+      "trimType": "auto",
+      "trimThreshold": 1,
+      "rotated": false,
+      "offsetX": 0,
+      "offsetY": 0,
+      "trimX": 0,
+      "trimY": 0,
+      "width": 1280,
+      "height": 922,
+      "rawWidth": 1280,
+      "rawHeight": 922,
+      "borderTop": 0,
+      "borderBottom": 0,
+      "borderLeft": 0,
+      "borderRight": 0,
+      "subMetas": {}
+    }
+  }
+}
